+++ quickstarts/maven/spring-boot-camel-complete/src/main/java/org/eclipse/jkube/quickstart/springboot/camel/log/CamelServletConfiguration.java
@@ -1,0 +1,30 @@
+/**
+ * Copyright (c) 2019 Red Hat, Inc.
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at:
+ *
+ *     https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   Red Hat, Inc. - initial API and implementation
+ */
+package org.eclipse.jkube.quickstart.springboot.camel.log;
+
+import org.apache.camel.component.servlet.CamelHttpTransportServlet;
+import org.springframework.boot.web.servlet.ServletRegistrationBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class CamelServletConfiguration {
+  @Bean
+  public ServletRegistrationBean<CamelHttpTransportServlet> camelServletRegistrationBean() {
+    final ServletRegistrationBean<CamelHttpTransportServlet> registration = new ServletRegistrationBean<>(
+        new CamelHttpTransportServlet(), "/camel/*");
+    registration.setName("CamelServlet");
+    return registration;
+  }
+}